+++ Changelog.md
@@ -11,6 +11,7 @@ __New Features__
 
 __Bugfixes__
 - Adds error-checking for `NumberofConditionedFloorsAboveGrade`=0, which is not allowed per the documentation.
+- Fixes error if a heating system and a heat pump (that only provides cooling) are attached to the same distribution system.
 - Misc Manual J design load calculation improvements.
 
 ## OpenStudio-ERI v1.8.0